Refactor HistoryColumn for another class use it.
